Move core.py publisher code into shared
