+++ stream_alert/alert_processor/helpers.py
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-from publishers.core import AlertPublisherRepository
+from stream_alert.shared.publisher import AlertPublisherRepository
 
 
 def elide_string_middle(text, max_length):
